Try a slightly different approach.
Convert the sub structs into pointers in their interface which allows
us slightly simpler logic inside Unmarshal.

We use a slightly weird way to access the map because we know it
should only have one key (this is checked), so `for v := range m` is
used to select the only key in the map.