A different approach again.
For unmarshal we create a name of type names and their reflect.Type.
We use that to allocate a new value of that type to create the
value we decode into.

For marshal, we can do it even simpler (I considered parsing the
string returned by reflect.Type.String(), but that was too horrible)
just use the tface interface to provide us a function that gives us
a type name.